backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
boto3==1.35.74
python-dotenv==1.0.1
anthropic==0.39.0
reportlab==4.0.7
pillow==10.0.0
xhtml2pdf==0.2.16
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
